+++ spring-batch-core/src/main/java/org/springframework/batch/core/step/skip/AlwaysSkipItemSkipPolicy.java
@@ -26,7 +26,7 @@ package org.springframework.batch.core.step.skip;
 public class AlwaysSkipItemSkipPolicy implements SkipPolicy {
 
 	@Override
-	public boolean shouldSkip(Throwable t, int skipCount) {
+	public boolean shouldSkip(Throwable t, long skipCount) {
 		return true;
 	}
